+++ R/wedge.R
@@ -1,9 +1,10 @@
 #' Performs the wedge operation
 #'
 #' Calculates the result from the wedge operation on two real coefficient
-#' matrices, two complex coefficient vectors or two phylo objects.
+#' matrices, two y evaluated polynomial coefficient vectors or two phylo objects.
 #'
 #' @param A,B two real coefficient matrices, complex coefficient vectors or phylo objects
+#' @inheritParams treeToPoly
 #' @return the wedge result in the same form as the arguments
 #' @import ape
 #' @useDynLib treenomial
@@ -21,19 +22,30 @@
 #' # wedge two complex coefficient vectors
 #'
 #' leaf <- as.complex(c(0,1))
-#' wedge(leaf, leaf)
+#' wedge(leaf, leaf, "yEvaluated",5)
 #'
 #' @export
-wedge <- function(A, B) {
+wedge <- function(A, B, type = c("default","yEvaluated","phylo"), y) {
 
-  type = class(A[1])
+  if(!missing(type) & length(type)>1) stop("only one 'type' allowed")
+  type <- match.arg(type)
 
-  if (type == "numeric" && class(A) == "matrix") {
+  if(type == "default"){
+    if(typeof(A) != typeof(B)){
+      stop("wedge operands type must match")
+    }
+
     wedgeExport(A,B)
-  } else if (type == "complex") {
-    as.vector(wedgeExportConv(A,B))
-  } else {
-    if(all(A == "leaf" & B == "leaf")){
+
+  } else if(type == "yEvaluated"){
+    if(typeof(A) != typeof(B)){
+      stop("wedge operands type must match")
+    }
+
+    as.vector(wedgeExportConv(A,B,y))
+
+  } else if(type == "phylo"){
+    if(all(c(A == "leaf", B == "leaf"))){
       res <- rtree(2, br = NULL)
     } else if (is.character(A)){
       res <- bind.tree(rtree(2, br = NULL), B, where = 1)
@@ -52,5 +64,5 @@ wedge <- function(A, B) {
     }
     res$tip.label <- 1:(length(res$tip.label))
     return(res)
-  }
+ }
 }