+++ tests/testthat/test-wedge.R
@@ -15,18 +15,18 @@ test_that("Check wedging result real poly", {
 test_that("Check wedging result complex poly", {
 
   leaf <- as.complex(c(0,1))
-  cherry <- wedge(leaf, leaf)
+  cherry <- wedge(leaf, leaf, type = "yEvaluated", y =(1+1i))
 
   cherryWedged <- c(1+3i,0,2i+2,0,1)
 
-  expect_equal(wedge(cherry, cherry), cherryWedged)
+  expect_equal(wedge(cherry, cherry, type = "yEvaluated", y = (1+1i)), cherryWedged)
 })
 
 test_that("Check wedging result phylo type", {
 
   res <- allTrees(2, type = "phylo")[[2]]
 
-  expect_equal(wedge("leaf", "leaf"), res[[1]])
+  expect_equal(wedge("leaf", "leaf", type = "phylo"), res[[1]])
 })